Split test of config in separate file
